+++ crates/forge/src/lib.rs
@@ -9,7 +9,8 @@ use forge_runner::forge_config::ForgeTrackedResource;
 use foundry_ui::components::warning::WarningMessage;
 use foundry_ui::{Message, UI};
 use run_tests::workspace::run_for_workspace;
-use scarb_api::{ScarbCommand, metadata::MetadataCommandExt};
+use scarb_api::ScarbCommand;
+use scarb_api::metadata::metadata;
 use scarb_ui::args::{FeaturesSpec, PackagesFilter, ProfileSpec};
 use semver::Version;
 use shared::auto_completions::{Completions, generate_completions};
@@ -18,7 +19,6 @@ use std::ffi::OsString;
 use std::sync::Arc;
 use std::{fs, num::NonZeroU32, thread::available_parallelism};
 use tokio::runtime::Builder;
-use universal_sierra_compiler_api::UniversalSierraCompilerCommand;
 
 pub mod block_number_map;
 mod clean;
@@ -35,7 +35,7 @@ mod warn;
 pub const CAIRO_EDITION: &str = "2024_07";
 
 const MINIMAL_SCARB_VERSION: Version = Version::new(2, 10, 0);
-const MINIMAL_RECOMMENDED_SCARB_VERSION: Version = Version::new(2, 10, 1);
+const MINIMAL_RECOMMENDED_SCARB_VERSION: Version = Version::new(2, 11, 4);
 const MINIMAL_USC_VERSION: Version = Version::new(2, 0, 0);
 const MINIMAL_SNFORGE_STD_VERSION: Version = Version::new(0, 50, 0);
 const MINIMAL_SNFORGE_STD_DEPRECATED_VERSION: Version = Version::new(0, 50, 0);
@@ -303,7 +303,7 @@ pub fn main_execution(ui: Arc<UI>) -> Result<ExitStatus> {
         }
         ForgeSubcommand::CleanCache {} => {
             ui.println(&WarningMessage::new("`snforge clean-cache` is deprecated and will be removed in the future. Use `snforge clean cache` instead"));
-            let scarb_metadata = ScarbCommand::metadata().inherit_stderr().run()?;
+            let scarb_metadata = metadata()?;
             let cache_dir = scarb_metadata.workspace.root.join(CACHE_DIR);
 
             if cache_dir.exists() {
@@ -366,7 +366,7 @@ fn check_requirements(output_on_success: bool, ui: &UI) -> Result<()> {
 
     requirements_checker.add_requirement(Requirement {
         name: "Universal Sierra Compiler".to_string(),
-        command: RefCell::new(UniversalSierraCompilerCommand::new().arg("--version").command()),
+        command: RefCell::new(universal_sierra_compiler_api::version_command()?),
         minimal_version: MINIMAL_USC_VERSION,
         minimal_recommended_version: None,
         helper_text: "Reinstall `snforge` using the same installation method or follow instructions from https://foundry-rs.github.io/starknet-foundry/getting-started/installation.html#universal-sierra-compiler-update".to_string(),